feature: the Target now avoids the Player
The target now won't bounce in the direction of the player. For that, it
uses the functions added in the previous patch and adds the function
getRandomDirectionAvoiding(object) to Target. (if you have a better idea
for a name, tell me :D). Also, target now has a reference to player to
avoid finding it every update amongst the gameobjects...