style & feature: cleaned up and improved score
Cleaned up the code that calculated the score and highscore, things like
where it was calculated or simple style changes.
Removed scoreChanged signal because taragetReached was more appropriate
and because it would overlap with the new score variable and its
onScoreChanged signal.
Also added a new Timer that calculates the score based on the number of
points and the time once per second. It will be used to display the
score in-game and having it change in real time.