fix: lazer instant-killing player
The long-awaited bug fix is here! A short explanation of the bug:
- the player touches the target
- TargetReached is emitted
- (sometimes respawn)
- lazer is generated at a some unchecked position
- (sometimes respawn)
You see? Sometimes the Lazer would spawn (at a totally random position)
after things had been spawn, thus not placing it at a good position.
The fix: when generating the Lazer, respawn it immediately so there's no
possibility of bug.